Use threaded streamer
