--- src/cpp/src/whisper/streamer.cpp
@@ -1,13 +1,0 @@
-// Copyright (C) 2023-2025 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-
-#include "streamer.hpp"
-
-#include "text_callback_streamer.hpp"
-
-namespace ov {
-namespace genai {
-
-
-}  // namespace genai
-}  // namespace ov
